# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: parental
    description: Parent safety preferences
  - name: scheduled-messages
    description: Authenticated scheduled chat message operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /parental/alert-preferences:
    get:
      operationId: getParentAlertPreferences
      tags: [parental]
      summary: Get the authenticated parent's content alert threshold
      responses:
        "200":
          description: Current alert threshold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertPreference"
        "401":
          description: Invalid or missing auth token
    put:
      operationId: updateParentAlertPreferences
      tags: [parental]
      summary: Update the authenticated parent's content alert threshold
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AlertPreferenceUpdate"
      responses:
        "200":
          description: Updated alert threshold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertPreference"
        "400":
          description: Invalid alert threshold
        "401":
          description: Invalid or missing auth token
  /scheduled-messages:
    get:
      operationId: listScheduledMessages
      tags: [scheduled-messages]
      summary: List the authenticated sender's pending scheduled messages
      responses:
        "200":
          description: Pending scheduled messages
          content:
            application/json:
              schema:
                type: object
                required: [scheduledMessages]
                properties:
                  scheduledMessages:
                    type: array
                    items:
                      $ref: "#/components/schemas/ScheduledMessage"
        "401":
          description: Invalid or missing auth token
    post:
      operationId: createScheduledMessage
      tags: [scheduled-messages]
      summary: Schedule a text message in a direct or group chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduledMessageInput"
      responses:
        "201":
          description: Scheduled message created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledMessage"
        "400":
          description: Invalid message or scheduled time
        "401":
          description: Invalid or missing auth token
        "403":
          description: Sender is not allowed to message this chat
        "404":
          description: Chat not found
  /scheduled-messages/{scheduledMessageId}:
    patch:
      operationId: updateScheduledMessage
      tags: [scheduled-messages]
      summary: Edit a pending scheduled message owned by the authenticated sender
      parameters:
        - $ref: "#/components/parameters/ScheduledMessageId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduledMessageUpdate"
      responses:
        "200":
          description: Scheduled message updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduledMessage"
        "400":
          description: Invalid message or scheduled time
        "401":
          description: Invalid or missing auth token
        "404":
          description: Pending scheduled message not found
        "409":
          description: Message is already being sent
    delete:
      operationId: cancelScheduledMessage
      tags: [scheduled-messages]
      summary: Cancel a pending scheduled message owned by the authenticated sender
      parameters:
        - $ref: "#/components/parameters/ScheduledMessageId"
      responses:
        "204":
          description: Scheduled message cancelled
        "401":
          description: Invalid or missing auth token
        "404":
          description: Pending scheduled message not found
components:
  parameters:
    ScheduledMessageId:
      name: scheduledMessageId
      in: path
      required: true
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AlertPreference:
      type: object
      properties:
        ok:
          type: boolean
        minimumSeverity:
          $ref: "#/components/schemas/AlertThreshold"
      required:
        - minimumSeverity
    AlertPreferenceUpdate:
      type: object
      properties:
        minimumSeverity:
          $ref: "#/components/schemas/AlertThreshold"
      required:
        - minimumSeverity
    AlertThreshold:
      type: string
      enum: [all, medium, high]
    ScheduledMessage:
      type: object
      required: [id, chatId, chatName, text, scheduledFor, status, createdAt, updatedAt]
      properties:
        id:
          type: string
        chatId:
          type: string
        chatName:
          type: string
        text:
          type: string
        scheduledFor:
          type: integer
          format: int64
        status:
          type: string
          enum: [pending, sending, sent, cancelled, failed]
        createdAt:
          type: integer
          format: int64
        updatedAt:
          type: integer
          format: int64
        sentMessageId:
          type: [string, "null"]
        failureReason:
          type: [string, "null"]
    ScheduledMessageInput:
      type: object
      required: [chatId, text, scheduledFor]
      properties:
        chatId:
          type: string
        text:
          type: string
          minLength: 1
          maxLength: 2000
        scheduledFor:
          type: integer
          format: int64
    ScheduledMessageUpdate:
      type: object
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 2000
        scheduledFor:
          type: integer
          format: int64