# lib/api-zod/src/generated/types/scheduledMessage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScheduledMessageStatus } from "./scheduledMessageStatus";

export interface ScheduledMessage {
  id: string;
  chatId: string;
  chatName: string;
  text: string;
  scheduledFor: number;
  status: ScheduledMessageStatus;
  createdAt: number;
  updatedAt: number;
  /** @nullable */
  sentMessageId?: string | null;
  /** @nullable */
  failureReason?: string | null;
}