# lib/api-zod/src/generated/types/scheduledMessageInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduledMessageInput {
  chatId: string;
  /**
   * @minLength 1
   * @maxLength 2000
   */
  text: string;
  scheduledFor: number;
}