# lib/api-zod/src/generated/types/scheduledMessageUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ScheduledMessageUpdate {
  /**
   * @minLength 1
   * @maxLength 2000
   */
  text?: string;
  scheduledFor?: number;
}